requirements.txt
# Heroku Deployment Files

## 1. requirements.txt
python-telegram-bot>=20.0
httpx
sqlite3
python-dotenv

## 2. Procfile
worker: python bot.py

## 3. runtime.txt
python-3.11.7

## 4. .gitignore
*.pyc
__pycache__/
.env
*.db
venv/
.DS_Store
bot.log
documents/
photos/